TextBox slightly less broken, but still broken as fuck + added "press and hold" for right left and backspace
if your press the Right, Left or Back buttons for more then 2.5 seconds,
it will start Automatic mode.

the backspace button is still broken, but most of the other stuff are
fixed now.

also added console.beepS for when you're trying to move the Tab beyond
the box's borders.
It's pretty neat.